% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_cherry}
\alias{as_cherry}
\title{Cherry class constructor}
\usage{
as_cherry(x, tree, game_diff)
}
\arguments{
\item{x}{An integer value that represents total number of spins.}

\item{tree}{Size of the original tree.}

\item{game_diff}{Progression of the game based on the sequential differences
in the amount of cherries on the tree.}
}
\value{
An object of class cherry
}
\description{
Cherry class constructor
}
