% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_game.R
\name{play_game}
\alias{play_game}
\title{Function to play hi-ho cherry-o!}
\usage{
play_game(tree, max_game = 1e+05)
}
\arguments{
\item{tree}{A integer value indicating the size of the tree.}

\item{max_game}{The maximum number of spins that are allowed before the
game suspends.}
}
\value{
An object of class 'cherry' which stores the number of spins
in the game along with the progression of the game stored as an attribute.
}
\description{
This function simulates hi-ho cherry-o using the standard spinner but
allow the user to change the size of the cherry tree.
}
