% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generic.R
\name{plot.cherry}
\alias{plot.cherry}
\title{Plot the progression of a high-ho cherry-o game.}
\usage{
\method{plot}{cherry}(x, y, ...)
}
\arguments{
\item{x}{An object of class 'cherry'}

\item{y}{Ignored}

\item{...}{Additional arguments to \code{\link[=plot.default]{plot.default()}}.}
}
\description{
Plot the progression of a high-ho cherry-o game.
}
