% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_game.R
\name{spinner}
\alias{spinner}
\title{The spinner in the game hi-ho cherry-o.}
\usage{
spinner(tree, basket)
}
\arguments{
\item{tree}{An integer value specifying the current number of cherries
on the tree.}

\item{basket}{An integer value specifying the current number of cherries
in the basket.}
}
\value{
An integer value specifying the number of cherries left on the tree.
}
\description{
The spinner in the game hi-ho cherry-o.
}
